Custom camera controls.
